add functions for creating observables
