+++ pennylane/hf/__init__.py
@@ -17,12 +17,11 @@ Hartree-Fock calculations.
 """
 from .basis_data import STO3G, atomic_numbers
 from .basis_set import BasisFunction, atom_basis_data, mol_basis_data
-from .dipole import dipole_integrals, dipole_moment, fermionic_dipole, fermionic_one, qubit_operator
+from .dipole import dipole_integrals, dipole_moment, fermionic_dipole
 from .hamiltonian import (
     generate_electron_integrals,
     generate_fermionic_hamiltonian,
     generate_hamiltonian,
-    simplify,
 )
 from .hartree_fock import generate_scf, hf_energy, nuclear_energy
 from .integrals import (
@@ -62,3 +61,4 @@ from .tapering import (
     transform_hamiltonian,
     transform_hf,
 )
+from .transform import fermionic_operator, jordan_wigner, qubit_operator
