+++ pennylane/hf/hamiltonian.py
@@ -19,6 +19,7 @@ import autograd.numpy as anp
 import pennylane as qml
 from pennylane import numpy as np
 from pennylane.hf.hartree_fock import generate_scf, nuclear_energy
+from pennylane.hf.transform import fermionic_operator, qubit_operator
 
 
 def generate_electron_integrals(mol, core=None, active=None):
@@ -173,33 +174,7 @@ def generate_fermionic_hamiltonian(mol, cutoff=1.0e-12, core=None, active=None):
         """
         core_constant, one, two = generate_electron_integrals(mol, core, active)(*args)
 
-        core_constant = anp.array([core_constant])
-
-        indices_one = anp.argwhere(abs(one) >= cutoff)
-        operators_one = (indices_one * 2).tolist() + (
-            indices_one * 2 + 1
-        ).tolist()  # up-up + down-down terms
-        coeffs_one = anp.tile(one[abs(one) >= cutoff], 2)
-
-        indices_two = anp.argwhere(abs(two) >= cutoff)
-        n = len(indices_two)
-        operators_two = (
-            [(indices_two[i] * 2).tolist() for i in range(n)]  # up-up-up-up term
-            + [
-                (indices_two[i] * 2 + [0, 1, 1, 0]).tolist() for i in range(n)
-            ]  # up-down-down-up term
-            + [
-                (indices_two[i] * 2 + [1, 0, 0, 1]).tolist() for i in range(n)
-            ]  # down-up-up-down term
-            + [(indices_two[i] * 2 + 1).tolist() for i in range(n)]  # down-down-down-down term
-        )
-        coeffs_two = anp.tile(two[abs(two) >= cutoff], 4) / 2
-
-        coeffs = anp.concatenate((core_constant, coeffs_one, coeffs_two))
-        operators = [[]] + operators_one + operators_two
-        indices_sort = [operators.index(i) for i in sorted(operators)]
-
-        return coeffs[indices_sort], sorted(operators)
+        return fermionic_operator(core_constant, one, two, cutoff)
 
     return fermionic_hamiltonian
 
@@ -242,270 +217,6 @@ def generate_hamiltonian(mol, cutoff=1.0e-12, core=None, active=None):
         """
         h_ferm = generate_fermionic_hamiltonian(mol, cutoff, core, active)(*args)
 
-        ops = []
-
-        for n, t in enumerate(h_ferm[1]):
-
-            if len(t) == 0:
-                coeffs = anp.array([0.0])
-                coeffs = coeffs + np.array([h_ferm[0][n]])
-                ops = ops + [qml.Identity(0)]
-
-            elif len(t) == 2:
-                op = _generate_qubit_operator(t)
-                if op != 0:
-                    for i, o in enumerate(op[1]):
-                        if len(o) == 0:
-                            op[1][i] = qml.Identity(0)
-                        if len(o) == 1:
-                            op[1][i] = _return_pauli(o[0][1])(o[0][0])
-                        if len(o) > 1:
-                            k = qml.Identity(0)
-                            for o_ in o:
-                                k = k @ _return_pauli(o_[1])(o_[0])
-                            op[1][i] = k
-                    coeffs = np.concatenate([coeffs, np.array(op[0]) * h_ferm[0][n]])
-                    ops = ops + op[1]
-
-            elif len(t) == 4:
-                op = _generate_qubit_operator(t)
-                if op != 0:
-                    for i, o in enumerate(op[1]):
-                        if len(o) == 0:
-                            op[1][i] = qml.Identity(0)
-                        if len(o) == 1:
-                            op[1][i] = _return_pauli(o[0][1])(o[0][0])
-                        if len(o) > 1:
-                            k = qml.Identity(0)
-                            for o_ in o:
-                                k = k @ _return_pauli(o_[1])(o_[0])
-                            op[1][i] = k
-                    coeffs = np.concatenate([coeffs, np.array(op[0]) * h_ferm[0][n]])
-                    ops = ops + op[1]
-
-        h = simplify(qml.Hamiltonian(coeffs, ops), cutoff=cutoff)
-
-        return h
+        return qubit_operator(h_ferm)
 
     return hamiltonian
-
-
-def _generate_qubit_operator(op):
-    r"""Convert a fermionic operator to a qubit operator using the Jordan-Wigner mapping.
-
-    The one-body fermionic operator :math:`a_2^\dagger a_0` is constructed as [2, 0] and the
-    two-body operator :math:`a_4^\dagger a_3^\dagger a_2 a_1` is constructed as [4, 3, 2, 1].
-
-    Args:
-        op (list[int]): the fermionic operator
-
-    Returns
-        tuple(list[complex], list[list[int, str]]): list of coefficients and the qubit-operator terms
-
-    **Example**
-
-    >>> f  = [0, 0]
-    >>> q = _generate_qubit_operator(f)
-    >>> q
-    ([(0.5+0j), (-0.5+0j)], [[], [(0, 'Z')]]) # corresponds to :math:`\frac{1}{2}(I_0 - Z_0)`
-    """
-    if len(op) == 1:
-        op = [((op[0], 1),)]
-
-    if len(op) == 2:
-        op = [((op[0], 1), (op[1], 0))]
-
-    if len(op) == 4:
-        op = [((op[0], 1), (op[1], 1), (op[2], 0), (op[3], 0))]
-
-        if op[0][0][0] == op[0][1][0] or op[0][2][0] == op[0][3][0]:
-            return 0
-
-    for t in op:
-        for l in t:
-            z = [(index, "Z") for index in range(l[0])]
-            x = z + [(l[0], "X"), 0.5]
-
-            if l[1]:
-                y = z + [(l[0], "Y"), -0.5j]
-
-            else:
-                y = z + [(l[0], "Y"), 0.5j]
-
-            if t.index(l) == 0:
-                q = [x, y]
-            else:
-                m = []
-                for t1 in q:
-                    for t2 in [x, y]:
-                        q1, c1 = _pauli_mult(t1[:-1], t2[:-1])
-                        m.append(q1 + [c1 * t1[-1] * t2[-1]])
-                q = m
-
-    c = [p[-1] for p in q]
-    o = [p[:-1] for p in q]
-
-    for item in o:
-        k = [i for i, x in enumerate(o) if x == item]
-        if len(k) >= 2:
-            for j in k[::-1][:-1]:
-                del o[j]
-                c[k[0]] = c[k[0]] + c[j]
-                del c[j]
-
-    return c, o
-
-
-def simplify(h, cutoff=1.0e-12):
-    r"""Add together identical terms in the Hamiltonian.
-
-    The Hamiltonian terms with identical Pauli words are added together and eliminated if the
-    overall coefficient is smaller than a cutoff value.
-
-    Args:
-        h (Hamiltonian): PennyLane Hamiltonian
-        cutoff (float): cutoff value for discarding the negligible terms
-
-    Returns:
-        Hamiltonian: Simplified PennyLane Hamiltonian
-
-    **Example**
-
-    >>> c = np.array([0.5, 0.5])
-    >>> h = qml.Hamiltonian(c, [qml.PauliX(0) @ qml.PauliY(1), qml.PauliX(0) @ qml.PauliY(1)])
-    >>> print(simplify(h))
-    (1.0) [X0 Y1]
-    """
-    wiremap = dict(zip(h.wires, range(len(h.wires) + 1)))
-
-    c = []
-    o = []
-    for i, op in enumerate(h.terms()[1]):
-        op = qml.operation.Tensor(op).prune()
-        op = qml.grouping.pauli_word_to_string(op, wire_map=wiremap)
-        if op not in o:
-            c.append(h.terms()[0][i])
-            o.append(op)
-        else:
-            c[o.index(op)] += h.terms()[0][i]
-
-    coeffs = []
-    ops = []
-    nonzero_ind = np.argwhere(abs(np.array(c)) > cutoff).flatten()
-    for i in nonzero_ind:
-        coeffs.append(c[i])
-        ops.append(qml.grouping.string_to_pauli_word(o[i], wire_map=wiremap))
-
-    try:
-        coeffs = qml.math.stack(coeffs)
-    except ValueError:
-        pass
-
-    return qml.Hamiltonian(coeffs, ops)
-
-
-def _pauli_mult(p1, p2):
-    r"""Return the result of multiplication between two tensor products of Pauli operators.
-
-    The Pauli operator :math:`(P_0)` is denoted by [(0, 'P')], where :math:`P` represents
-    :math:`X`, :math:`Y` or :math:`Z`.
-
-    Args:
-        p1 (list[tuple[int, str]]): the first tensor product of Pauli operators
-        p2 (list[tuple[int, str]]): the second tensor product of Pauli operators
-
-    Returns
-        tuple(list[tuple[int, str]], complex): list of the Pauli operators and the coefficient
-
-    **Example**
-
-    >>> p1 = [(0, "X"), (1, "Y")],  # X_0 @ Y_1
-    >>> p2 = [(0, "X"), (2, "Y")],  # X_0 @ Y_2
-    >>> _pauli_mult(p1, p2)
-    ([(2, "Y"), (1, "Y")], 1.0) # p1 @ p2 = X_0 @ Y_1 @ X_0 @ Y_2
-    """
-    c = 1.0
-
-    t1 = [t[0] for t in p1]
-    t2 = [t[0] for t in p2]
-
-    k = []
-
-    for i in p1:
-        if i[0] in t1 and i[0] not in t2:
-            k.append((i[0], pauli_mult[i[1]]))
-        for j in p2:
-            if j[0] in t2 and j[0] not in t1:
-                k.append((j[0], pauli_mult[j[1]]))
-
-            if i[0] == j[0]:
-                if i[1] + j[1] in pauli_coeff:
-                    k.append((i[0], pauli_mult[i[1] + j[1]]))
-                    c = c * pauli_coeff[i[1] + j[1]]
-                else:
-                    k.append((i[0], pauli_mult[i[1] + j[1]]))
-
-    k = [i for i in k if "I" not in i[1]]
-
-    for item in k:
-        k_ = [i for i, x in enumerate(k) if x == item]
-        if len(k_) >= 2:
-            for j in k_[::-1][:-1]:
-                del k[j]
-
-    return k, c
-
-
-def _return_pauli(p):
-    r"""Return the PennyLane Pauli operator.
-
-    Args:
-        args (str): symbol representing the Pauli operator
-
-    Returns:
-        pennylane.ops: the PennyLane Pauli operator
-
-    **Example**
-
-    >>> _return_pauli('X')
-    qml.PauliX
-    """
-    if p == "X":
-        return qml.PauliX
-    if p == "Y":
-        return qml.PauliY
-
-    return qml.PauliZ
-
-
-pauli_mult = {
-    "XX": "I",
-    "YY": "I",
-    "ZZ": "I",
-    "ZX": "Y",
-    "XZ": "Y",
-    "ZY": "X",
-    "YZ": "X",
-    "XY": "Z",
-    "YX": "Z",
-    "IX": "X",
-    "IY": "Y",
-    "IZ": "Z",
-    "XI": "X",
-    "YI": "Y",
-    "ZI": "Z",
-    "I": "I",
-    "II": "I",
-    "X": "X",
-    "Y": "Y",
-    "Z": "Z",
-}
-
-pauli_coeff = {
-    "ZX": 1.0j,
-    "XZ": -1.0j,
-    "ZY": -1.0j,
-    "YZ": 1.0j,
-    "XY": 1.0j,
-    "YX": -1.0j,
-}
