+++ pennylane/hf/hartree_fock.py
@@ -214,7 +214,7 @@ def nuclear_energy(charges, r):
             coor = args[0]
         else:
             coor = r
-        e = 0
+        e = anp.array([0.0])
         for i, r1 in enumerate(coor):
             for j, r2 in enumerate(coor[i + 1 :]):
                 e = e + (charges[i] * charges[i + j + 1] / anp.linalg.norm(r1 - r2))
